pyproject.toml
[project]
name = "notion-mcp"
version = "0.1.0"
description = ""
authors = [
    {name = "hsm8867",email = "hsm4317@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.52.2,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-genai (>=1.57.0,<2.0.0)",
    "mcp (>=1.25.0,<2.0.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
